tech(TTB-060): Moved Options component to individual directory
