+++ client/src/App.jsx
@@ -3,7 +3,7 @@ import { BrowserRouter, Route, Routes } from 'react-router-dom';
 import Game from './game/Game';
 import GameRobot from './game/GameRobot';
 import Home from './components/Home/Home';
-import Options from './options/Options';
+import Options from './components/Options/Options';
 import Header from './components/Header/Header';
 import './app.css';
